requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=6.0.0,<7.0.0
pulumi-kubernetes>=4.0.0,<5.0.0
pulumi-eks>=2.0.0,<3.0.0
pyyaml>=6.0.0

